Fixes #5 - Now works on Vagrant 1.4.3
To be honest, I'm not super sure how these changes fix the issue? I just took what was the current head of `vagrant-omnibus` and tweaked it to Puppet. The main culprit is probably the double `call(env)` in install puppet that is removed by this commit. But, it works, so yay! :+1: